+++ src/torch_pointcloud/models/pointmlp.py
@@ -577,9 +577,9 @@ class PointMLPClassification(ClassificationModel):
 
     def configure_head(self) -> nn.Module:
         if self.num_classes == 0:
-            return nn.Identity()
+            return nn.Identity().train(self.training)
         if not self.head_channels:
-            return nn.Linear(self.num_features, self.num_classes)
+            return nn.Linear(self.num_features, self.num_classes).train(self.training)
         return MLP(
             [self.num_features] + list(self.head_channels) + [self.num_classes],
             act=self.act,
@@ -590,7 +590,7 @@ class PointMLPClassification(ClassificationModel):
             bias=True,
             dropout=self.head_dropout,
             plain_last=True,
-        )
+        ).train(self.training)
 
     def reset_classifier(self, num_classes: int, global_pool: Optional[PoolLike] = None, **kwargs: Any) -> None:
         self.num_classes = num_classes
@@ -758,8 +758,8 @@ class PointMLPSegmentation(SegmentationModel):
 
     def configure_head(self) -> nn.Module:
         if self.num_classes == 0:
-            return nn.Identity()
-        return nn.Linear(self.num_features, self.num_classes)
+            return nn.Identity().train(self.training)
+        return nn.Linear(self.num_features, self.num_classes).train(self.training)
 
     def reset_classifier(self, num_classes: int, **kwargs: Any) -> None:
         self.num_classes = num_classes