+++ src/torch_pointcloud/models/pointnext.py
@@ -589,9 +589,9 @@ class PointNeXtPartSegmentation(SegmentationModel):
 
     def configure_head(self) -> nn.Module:
         if self.num_classes == 0:
-            return nn.Identity()
+            return nn.Identity().train(self.training)
         if not self.head_channels:
-            return nn.Linear(self.num_features, self.num_classes)
+            return nn.Linear(self.num_features, self.num_classes).train(self.training)
         return MLP(
             [self.num_features] + list(self.head_channels) + [self.num_classes],
             act=self.act,
@@ -602,7 +602,7 @@ class PointNeXtPartSegmentation(SegmentationModel):
             bias=self.bias,
             dropout=self.dropout,
             plain_last=True,
-        )
+        ).train(self.training)
 
     def reset_classifier(self, num_classes: int, **kwargs: Any) -> None:
         self.num_classes = num_classes
@@ -827,9 +827,9 @@ class PointNeXtClassification(ClassificationModel):
 
     def configure_head(self) -> nn.Module:
         if self.num_classes == 0:
-            return nn.Identity()
+            return nn.Identity().train(self.training)
         if not self.head_channels:
-            return nn.Linear(self.num_features, self.num_classes)
+            return nn.Linear(self.num_features, self.num_classes).train(self.training)
         return MLP(
             [self.num_features] + list(self.head_channels) + [self.num_classes],
             act=self.act,
@@ -840,7 +840,7 @@ class PointNeXtClassification(ClassificationModel):
             bias=self.bias,
             dropout=self.dropout,
             plain_last=True,
-        )
+        ).train(self.training)
 
     def reset_classifier(self, num_classes: int, global_pool: Optional[PoolLike] = None, **kwargs: Any) -> None:
         self.num_classes = num_classes
@@ -1053,9 +1053,9 @@ class PointNeXtSegmentation(SegmentationModel):
 
     def configure_head(self) -> nn.Module:
         if self.num_classes == 0:
-            return nn.Identity()
+            return nn.Identity().train(self.training)
         if not self.head_channels:
-            return nn.Linear(self.num_features, self.num_classes)
+            return nn.Linear(self.num_features, self.num_classes).train(self.training)
         return MLP(
             [self.num_features] + list(self.head_channels) + [self.num_classes],
             act=self.act,
@@ -1066,7 +1066,7 @@ class PointNeXtSegmentation(SegmentationModel):
             bias=self.bias,
             dropout=self.dropout,
             plain_last=True,
-        )
+        ).train(self.training)
 
     def reset_classifier(self, num_classes: int, **kwargs: Any) -> None:
         self.num_classes = num_classes