+++ src/torch_pointcloud/models/pointpillars.py
@@ -336,7 +336,7 @@ class PointPillarsDetection(DetectionModel):
             num_dir_bins=self.num_dir_bins,
             dir_offset=self.dir_offset,
             dir_limit_offset=self.dir_limit_offset,
-        )
+        ).train(self.training)
 
     def forward_features(self, voxels: Tensor, pos_voxel: Tensor, voxel_num_points: Tensor, batch: Tensor) -> Tensor:
         voxel_indices = torch.cat([batch.view(-1, 1).to(pos_voxel), pos_voxel], dim=1)
@@ -492,7 +492,7 @@ class PointPillarsMultiHeadDetection(DetectionModel):
             act_kwargs=self.act_kwargs,
             norm=self.norm,
             norm_kwargs=self.norm_kwargs,
-        )
+        ).train(self.training)
 
     def forward_features(self, voxels: Tensor, pos_voxel: Tensor, voxel_num_points: Tensor, batch: Tensor) -> Tensor:
         voxel_indices = torch.cat([batch.view(-1, 1).to(pos_voxel), pos_voxel], dim=1)